MAIL added support for weak pointers
Added MLton.Weak: MLTON_WEAK.

Added a new primitive unary type constructor, 'a weak.

Added the following primitives, which are eventually turned into a C
call, but need to be primitives because various analyses (flow
analysis, constant propagation, ...) need to handle the primitives
specially.

	MLton prim	GC C call
	-----------	-------------
	Weak_canGet	GC_weakCanGet
	Weak_get	GC_weakGet
	Weak_new	GC_weakNew

Weak_new allocates a new weak, Weak_get returns the weak pointer
(which may be bogus), and Weak_canGet tells whether the weak pointer
is bogus or not.

Extended the flow analysis, constant propagation, and useless passes
to handle weaks, essentially treating them like references.

SsaToRssa implements the weak primitives.  Representation analysis
determines if the weak type is a pointer, and if so, SsaToRssa inserts
the appropriate call to the C function.  Otherwise, SsaToRssa
implements Weak_canGet to always return false.

Added two new ObjectType.t to Rssa/Machine.

	     | Weak of Type.t (* in Weak t, must have Type.isPointer t *)
	     | WeakGone

Weak t is the type of weak objects where the pointer points to an
object of type t.  I had initially use "Weak of PointerTycon.t", but
that didn't work because it also makes sense for IntInfs to have weak
pointers to them.  WeakGone is used by the runtime system to indicate
a weak object whose pointer has been nulled.

Recall that ObjectType.t corresponds to the runtime GC_objectType, and
that there is a per-program array of GC_objectTypes that is indexed
into by object headers.  For each Weak t, the backend adds an entry to
the objectTypes array with WEAK_TAG, numPointers = 1, and
numNonPointers = 1.  It also arranges to put an objectType at the
WEAK_GONE_TYPE_INDEX with WEAK_TAG, numPointers = 0, and
numNonPointers = 2 -- this runtime replaces a weak object header with
WEAK_GONE_HEADER when a weak pointer is nulled.

Only the runtime knows how weaks are represented.  A 'a Weak.t is
represented as a GC object with a header and two words.  The header
indexes into an element of objectTypes with WEAK_TAG.  If the the weak
pointer is valid, then the numPointers = 1 and numNonPointers = 1.
Otherwise, it is the special WEAK_GONE_HEADER, and the type has
numPointers = 0 and numNonPointers = 2.  In a weak object, the first
word is a link pointer, which is only used during copying collection.
The second word is the weak pointer, or is BOGUS_POINTER if the weak
pointer is invalid.

The runtime is responsible for forwarding or nulling weak pointers,
depending on whether the object pointed to is retained by the GC.
Depending on the gc strategy, copying or mark-compact, this is done in
different ways.

For copying gc (major or minor), the copying phase treats a weak
object as an object with no pointers.  However, whenever a weak object
is forwarded, if the weak pointer points to an object in from space,
i.e. it points into the nursery or we are doing a major gc, then the
forwarded copy is linked into a linked list of all weak objects with
pointers that need to be updated.  After the forwarding is done,
updateWeaks () traverses the linked list and either updates the weak
pointer if it points to a forwarded object or nulls the weak pointer
otherwise.  For now, there can't be intergenerational weak pointers
since we don't have Weak.set.  If we ever add Weak.set, we will need
to have forwardInterGenerationalPointers link weak objects it finds
into the weak list.

Mark-compact GC is simpler.  The mark phase treats a weak object as an
object with no pointers.  Then, the updateForwardPointers phase clears
the weak pointer if it points to an unmarked object.  Otherwise, the
weak object is treated according to what the objectType says about the
numPointers.