MAIL support for multiple integer, real, and word sizes
Added compiler support for multiple integer, real, and word sizes.
This checkin just changes the compiler infrastructure.  It does not
add the basis library modules.  There are also some still some holes,
especially in the x86 codegen support.  Matthew, if you could take a
look at those, that would be great.

Everything that was there before is still supported, though, and all
regressions and self compiles pass.

The main new datatypes are:

	datatype IntSize.t = I8 | I16 | I32 | I64
	datatype RealSize.t = R32 | R64
	datatype WordSize.t = W8 | W16 | W32

Tycons and types have been extended so that they are parameterized
over the appropriate sizes.

      val int: IntSize.t -> t
      val real: RealSize.t -> t
      val word: WordSize.t -> t

There are also default values of each size (I32, R64, W32) and the
corresponding default types.

There are new modules for representing values of the various types at
runtime: IntX.t, RealX.t, and WordX.t.  The front end and ast
constants have been reworked to handle arbitrary size integers and
words.  Const has been reworked, adding variants for each of the
values, and replacing char with words of size W8 and string with word8
vector

      datatype Const.t =
	 Int of IntX.t
       | IntInf of IntInf.t
       | Real of RealX.t
       | Word of WordX.t
       | Word8Vector of Word8.t vector

I didn't treat IntInf as another size of Int because they are almost
always handled differently.

Integer, Real and Word primitives are now parameterized over the size
of value that they operate on.  The conversion operators have also
been generalized.

	datatype Prim.t =
		...
	     | Int_add of IntSize.t
		...
	     | Real_add of RealSize.t
		...
	     | Word_add of WordSize.t
		...
	     | Word_toIntX of WordSize.t * IntSize.t
		...

For such primitives, the prim name in the basis library now always has
the size as part of the name, e.g.

	_prim "Int32_addCheck": int * int -> int;

Case expressions in all the ILs have had cases on chars removed and
had cases on ints and words generalized to handle different sizes.

That's about it.  The plan now is to add support for Int8, Int16,
Int64, Real32, and Word16.  Hopefully, with all the infrastructure of
this checkin in place, adding those will only require writing some SML
basis library code and mucking with the codegen to implement the
primitives.  I would like Word64, but that will take a little more
work since there is no Word64 to bootstrap off of.

Feel free to start implementing the module of your choice, but please
announce it on the list first to avoid duplication of effort.