MAIL Updating Char.{fromString,scan}, String.{fromString,scan}
Updating these basis library functions according to my latest
understanding of the spec.  This is based on a discussion on the basis
library list in September 2003, which I concluded with the following
observations.

1. Char.scan returns NONE upon double quote.

2. String.scan stops upon double quote, and returns the output to that
   point.

3. String.scan consumes format sequences.

4. Char.scan consumes format sequences adjacent to a valid character.
   I mention this because of pathological character constants like
   #"a\ \", #"\ \a", #"\ \a\ \", and #"\ \\ \a".

5. Char.scan does not consume format sequences if it is unable to
   produce a character.  Consider the invalid character constant
   #"\ \".

6. String.scan never returns NONE.


I added some regressions to check these and updated our current
regressions for the changes in semantics (mostly due to String.scan
never returning NONE).