Added SML/NJ Library and CKit Library from SML/NJ (version 110.55) to distribution.
The libraries are distributed as a .tgz from the SML/NJ distribution
along with a patch file to adapt the libraries to MLton.

There is one relatively minor issue.  Both the SML/NJ HTML Library and
the CKit Library make use of MLLex lexers and MLYacc parsers.  Because
cm2mlb produces a very "tight" dependency tree, the minor differences
between SML/NJ's and MLton's mllex and mlyacc tools are relevant.
Likewise, differences with older versions of MLton's mllex would be
relevant.  Hence, I have included the derived files in the patch to
ensure compatibility.



SVN r3989