Minimize diff with SML/NJ upstream version of ml-yacc tool.
Since we now use the default int and word type provided by the host
compiler, there is no need for the extra 'type int = Int.int' and
':int' in the sources of ml-yacc.  Minimizing the diff with SML/NJ
upstream means that it is easier to apply changes.

SVN r7065